# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/funds/getRechargeRecords": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "获取充值记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "充值记录Id",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "外部标识",
                        "name": "externalIdentity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "充值Token",
                        "name": "token",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "创建开始时间戳",
                        "name": "createdStart",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "创建结束时间戳",
                        "name": "createdEnd",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页面大小",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pageIndex",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRechargeWalletResponse"
                        }
                    }
                }
            }
        },
        "/v1/funds/getRechargeWallet": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "获取充值钱包",
                "parameters": [
                    {
                        "description": " ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.GetRechargeWalletRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.GetRechargeWalletResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.GetRechargeWalletRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "callbackUrl": {
                    "type": "string"
                },
                "externalData": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "externalIdentity": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "controller.GetRechargeWalletResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "expireAt": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}